Roberta - PreLogin: aggiunto richiesta dei permessi; - NuovaPaginaTeoria: - ParametriAletrvista: set di paramentri da inviare ad altervista per il caricamneto del file - SaveOnaltervista: moficato per permettere la creazione di cartelle e salvataggio di file in esse -

+++ app/src/main/java/com/eschool/e_school/altervista/SaveOnAltervista.java
@@ -1,33 +1,53 @@
 package com.eschool.e_school.altervista;
 
 import android.os.AsyncTask;
+import android.support.v7.app.AlertDialog;
 import android.util.Log;
+import android.widget.Toast;
+
+import com.android.volley.Request;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.eschool.e_school.connessione.JsonRequest;
+import com.eschool.e_school.connessione.RequestSingleton;
 
 import org.apache.commons.net.ftp.FTP;
 import org.apache.commons.net.ftp.FTPClient;
+import org.apache.commons.net.ftp.FTPFile;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
 
 /**
  * Questa classe permette l'upload su altervista
  */
 
-public class SaveOnAltervista extends AsyncTask<File,Void,Void> {
+public class SaveOnAltervista extends AsyncTask<ParametriAltervista,Void,Void> {
+    ParametriAltervista param;
     File file;
+    String path;
+    Boolean risultato = false; // controllo sul risultato del caricamento
+
     @Override
     protected void onPostExecute(Void aVoid) {
         super.onPostExecute(aVoid);
         //una volta salvato lo elimino dal dispositivo
-        file.delete();
+        if (risultato)
+            file.delete();
+
     }
 
     @Override
-    protected Void doInBackground(File... files) {
+    protected Void doInBackground(ParametriAltervista... files) {
         FTPClient client = new FTPClient();
-        file = files[0];
+        param = files[0];
+        file = param.getF();
         try {
             client.connect("ftp.eschooldb.altervista.org");
         } catch (IOException e) {
@@ -59,9 +79,47 @@ public class SaveOnAltervista extends AsyncTask<File,Void,Void> {
         try {
             //Apposto di system.current ecc. metti il nome del file che ti interessa.
             //result = client.storeFile("/LOG/" + System.currentTimeMillis() +".pdf", in);
-            Log.d("DATO","name " + file.getName());
-            result = client.storeFile("/File/"+file.getName(), in);
+            Log.d("DATO", "name " + file.getName());
+            ArrayList<String> lista = new ArrayList<>();
+            FTPFile[] dirs = client.listDirectories("File");
+            for (FTPFile dir : dirs) {
+                Log.d("DATI", "nome dir " + dir.getName());
+                lista.add(dir.getName());
+            }
+
+            path = "File/" + param.getMateria();
+            if (!lista.contains(param.getMateria())) {
+                if (client.makeDirectory(path) && client.makeDirectory("File/" + param.getMateria() + "/" + param.getClasse())) {
+                    result = client.storeFile("File/" + param.getMateria() + "/" + param.getClasse() + "/" + file.getName(), in);
+                }
+            } else if (lista.contains(param.getMateria())) {
+                Log.d("DATI", "ENTRO!!");
+                //controllo l'esistenza della cartella relativa alla classe, se non esiste la creo, altrimenti carico il file
+                FTPFile[] dirMat = client.listDirectories(path);
+                for (FTPFile dir : dirMat) {
+                    Log.d("DATI", "nome dir " + dir.getName());
+                    lista.add(dir.getName());
+                }
+                if (!lista.contains(param.getClasse())) {
+                    if (client.makeDirectory(path + "/" + param.getClasse())) {
+                        result = client.storeFile(path + "/" + param.getClasse() + "/" + file.getName(), in);
+                    }
+                } else {
+                    FTPFile[] lFiles = client.listFiles(path + "/" + param.getClasse());
+                    for (FTPFile f : lFiles) {
+                        Log.d("DATI", "nome dir " + f.getName());
+                        lista.add(f.getName());
+                    }
+                    if (!lista.contains(file.getName()))
+                        result = client.storeFile(path + "/" + param.getClasse() + "/" + file.getName(), in);
+                    else
+                        Log.d("DATI", "Esiste");
+                }
+            } else
+                Log.d("DATI", "ERRORE");
+
         } catch (IOException e) {
+
             e.printStackTrace();
         }
         try {
@@ -69,7 +127,13 @@ public class SaveOnAltervista extends AsyncTask<File,Void,Void> {
         } catch (IOException e) {
             e.printStackTrace();
         }
-        if (result) Log.v("upload result", "succeeded");
+        if (result) {
+            risultato = true;
+
+            Log.v("upload result", "succeeded");
+        } else
+            Log.d("DATI", "errore");
+
         try {
             client.logout();
         } catch (IOException e) {
@@ -82,5 +146,23 @@ public class SaveOnAltervista extends AsyncTask<File,Void,Void> {
         }
         return null;
     }
-}
 
+    private void salvaFileDb() {
+        String url = "http://www.eschooldb.altervista.org/PHP/aggiungiFileTeoria.php";
+        HashMap<String, String> parametri = new HashMap<String, String>();
+        parametri.put("fileTeoria", path + "/" + param.getClasse() + "/" + file.getName());
+
+        JsonRequest richiesta = new JsonRequest(Request.Method.POST, url, parametri, new Response.Listener<JSONObject>() {
+            @Override
+            public void onResponse(JSONObject response) {
+
+            }
+        }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+
+            }
+        });
+        RequestSingleton.getInstance(param.getContext()).addToRequestQueue(richiesta);
+    }
+}
